Rename pkg nc to feedforward to reflect capabilities
