+++ java/de/uni_kl/cs/discodnc/DncTestConfig.java
@@ -29,7 +29,7 @@
 package de.uni_kl.cs.discodnc;
 
 import de.uni_kl.cs.discodnc.AlgDncBackend;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig;
 import de.uni_kl.cs.discodnc.numbers.NumBackend;
 
 import java.util.Set;
