+++ java/de/uni_kl/cs/discodnc/DncTestMethodSources.java
@@ -36,8 +36,8 @@ import de.uni_kl.cs.discodnc.curves.LinearSegment;
 import de.uni_kl.cs.discodnc.curves.disco.LinearSegment_Disco;
 import de.uni_kl.cs.discodnc.curves.disco.affine.Curve_Disco_Affine;
 import de.uni_kl.cs.discodnc.curves.disco.pwaffine.Curve_Disco_PwAffine;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig.ArrivalBoundMethod;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig.Multiplexing;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig.ArrivalBoundMethod;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig.Multiplexing;
 import de.uni_kl.cs.discodnc.numbers.NumBackend;
 
 import org.junit.jupiter.params.provider.Arguments;
