+++ java/de/uni_kl/cs/discodnc/TA_2S_1SC_4F_1AC_1P_Results.java
@@ -28,9 +28,9 @@
 
 package de.uni_kl.cs.discodnc;
 
-import de.uni_kl.cs.discodnc.nc.Analysis.Analyses;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig.ArrivalBoundMethod;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig.Multiplexing;
+import de.uni_kl.cs.discodnc.feedforward.Analysis.Analyses;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig.ArrivalBoundMethod;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig.Multiplexing;
 import de.uni_kl.cs.discodnc.numbers.Num;
 
 import java.util.Set;
